readme: m fx
'to see "the dump"' is an Ok phrase, but because of use of "the dump" as an
euphemism - such words are gaining stable subliminal messaging, which is well
known thing in linguistics, as all languages have a taboo words semantic shift drift
https://en.wikipedia.org/wiki/Semantic_change & frequent drift of naming & slang
especially well known around the implied theme.

The word "log" just suites as good & does not have such meanings.